feat: adding support for AWS Session Token as CLI input
This feature enhances the authentication functionality of
the AWS S3 Storage Backend for the Docker Registry.

In order to support role based access a AWS Session Token
has to be provided. The underlining package already has support for this
functionality but the startup configuration parameters of the
Docker Registry do not.